HOP-4172 : Forcing to a single thread in Beam doesn't work (fix, IT)
