+++ core/src/main/java/org/apache/hop/core/Const.java
@@ -283,6 +283,10 @@ public class Const {
   public static final String INTERNAL_VARIABLE_WORKFLOW_NAME =
       INTERNAL_VARIABLE_PREFIX + ".Workflow.Name";
 
+  /** The workflow ID */
+  public static final String INTERNAL_VARIABLE_WORKFLOW_ID =
+          INTERNAL_VARIABLE_PREFIX + ".Workflow.ID";
+
   /** The current pipeline directory */
   public static final String INTERNAL_VARIABLE_ENTRY_CURRENT_FOLDER =
       INTERNAL_VARIABLE_PREFIX + ".Entry.Current.Folder";
@@ -315,6 +319,10 @@ public class Const {
   public static final String INTERNAL_VARIABLE_PIPELINE_NAME =
       INTERNAL_VARIABLE_PREFIX + ".Pipeline.Name";
 
+  /** The pipeline ID */
+  public static final String INTERNAL_VARIABLE_PIPELINE_ID =
+          INTERNAL_VARIABLE_PREFIX + ".Pipeline.ID";
+
   /** The transform partition ID */
   public static final String INTERNAL_VARIABLE_TRANSFORM_PARTITION_ID =
       INTERNAL_VARIABLE_PREFIX + ".Transform.Partition.ID";
@@ -330,6 +338,12 @@ public class Const {
   /** The transform copy nr */
   public static final String INTERNAL_VARIABLE_TRANSFORM_COPYNR =
       INTERNAL_VARIABLE_PREFIX + ".Transform.CopyNr";
+
+  public static final String INTERNAL_VARIABLE_TRANSFORM_ID =
+          INTERNAL_VARIABLE_PREFIX + ".Transform.ID";
+
+  public static final String INTERNAL_VARIABLE_ACTION_ID =
+          INTERNAL_VARIABLE_PREFIX + ".Action.ID";
 
   /** The default maximum for the nr of lines in the GUI logs */
   public static final int MAX_NR_LOG_LINES = 5000;
