+++ engine/src/main/java/org/apache/hop/pipeline/engine/PipelineEngineFactory.java
@@ -157,6 +157,8 @@ public class PipelineEngineFactory {
 
     pipelineEngine.setPipelineMeta(pipelineMeta);
 
+    pipelineEngine.setInternalHopVariables(pipelineEngine);
+
     return pipelineEngine;
   }
 }
