+++ integration-tests/beam_directrunner/datasets/0005-generate-one-json-file-golden.csv
@@ -1,0 +1,40 @@
+First name,Last name,birthdate,city,cust_zip_code,housenr,id,state,stateCode,street
+llu-firstname,jdmdi-name,1974/07/01,pss-city," 18060"," 198",756,CALIFORNIA,CA,utrs-street
+eti-firstname,dhcim-name,1982/07/19,aed-city," 17140"," 156",765,FLORIDA,FL,syxk-street
+nnx-firstname,pvcck-name,1962/07/05,kfm-city," 19780"," 151",807,CALIFORNIA,CA,aybu-street
+xhp-firstname,gkiis-name,1954/01/02,bgh-city," 14440"," 53",269,FLORIDA,FL,btrx-street
+wpy-firstname,qxjcn-name,1954/12/19,rew-city," 11260"," 115",272,CALIFORNIA,CA,tldv-street
+egp-firstname,qrenr-name,1966/03/19,zwq-city," 16920"," 102",296,FLORIDA,FL,fbwi-street
+skq-firstname,cqapx-name,1958/09/02,akw-city," 11080"," 14",304,CALIFORNIA,CA,xakx-street
+sfu-firstname,ibfdt-name,1954/06/22,kqf-city," 11120"," 132",308,CALIFORNIA,CA,xbwg-street
+qtw-firstname,pwwep-name,1958/06/08,utn-city," 15160"," 176",320,FLORIDA,FL,sheh-street
+ofx-firstname,yxqsn-name,1974/02/27,ute-city," 18680"," 111",346,FLORIDA,FL,wqka-street
+wha-firstname,vnaov-name,1954/03/28,egm-city," 13120"," 20",12,CALIFORNIA,CA,hpep-street
+ayl-firstname,bwkoe-name,1978/07/16,rtw-city," 18660"," 179",25,CALIFORNIA,CA,mzww-street
+zio-firstname,rezku-name,1982/07/14,nvt-city," 19080"," 91",26,CALIFORNIA,CA,wwkd-street
+rgh-firstname,wzkjq-name,1974/08/12,hkm-city," 19000"," 134",44,CALIFORNIA,CA,yixf-street
+ayq-firstname,ieigg-name,1962/07/05,ljc-city," 12960"," 163",59,FLORIDA,FL,dnjz-street
+pvw-firstname,aavpj-name,1954/09/23,lai-city," 13780"," 171",70,FLORIDA,FL,nygu-street
+kzs-firstname,reeba-name,1966/07/08,zhc-city," 13100"," 88",96,FLORIDA,FL,abmv-street
+ovv-firstname,ejdoc-name,1954/11/15,ocs-city," 14920"," 169",125,FLORIDA,FL,bwhd-street
+niq-firstname,avxuw-name,1958/05/29,cyx-city," 13520"," 167",561,CALIFORNIA,CA,dxit-street
+ldq-firstname,mdftt-name,1966/09/03,mkm-city," 11460"," 9",600,CALIFORNIA,CA,rjkt-street
+tum-firstname,qxdyx-name,1958/10/21,mqw-city," 18100"," 156",389,FLORIDA,FL,gndl-street
+wlf-firstname,fmrwo-name,1962/05/03,qwb-city," 17820"," 22",422,CALIFORNIA,CA,stcz-street
+ttv-firstname,eqley-name,1974/09/16,trs-city," 12740"," 187",135,CALIFORNIA,CA,ibff-street
+ahc-firstname,nltvw-name,1958/03/16,uxf-city," 19520"," 131",177,CALIFORNIA,CA,fwsy-street
+cwd-firstname,ujfpc-name,1974/10/10,gki-city," 13800"," 12",178,FLORIDA,FL,sgiq-street
+kxv-firstname,bxerk-name,1982/02/18,sek-city," 14180"," 84",181,CALIFORNIA,CA,ctwu-street
+guv-firstname,rgryb-name,1966/05/06,xus-city," 11160"," 164",630,FLORIDA,FL,plka-street
+trc-firstname,temdh-name,1950/05/06,ald-city," 11720"," 1",662,FLORIDA,FL,tapb-street
+hzl-firstname,jikyx-name,1958/08/07,hgh-city," 13320"," 85",681,FLORIDA,FL,gmqp-street
+nmc-firstname,ymjve-name,1974/12/11,yux-city," 19280"," 65",897,FLORIDA,FL,ycqn-street
+etl-firstname,mpkrp-name,1958/06/02,kty-city," 10560"," 67",912,CALIFORNIA,CA,cpug-street
+oyu-firstname,papgo-name,1970/06/19,bet-city," 10280"," 155",919,CALIFORNIA,CA,otzt-street
+aaz-firstname,avrpe-name,1950/09/02,anm-city," 14000"," 197",202,FLORIDA,FL,ddjz-street
+nlv-firstname,nmccn-name,1974/07/05,dec-city," 11780"," 125",212,FLORIDA,FL,txyt-street
+vgc-firstname,yprgp-name,1971/01/05,alm-city," 14660"," 110",731,CALIFORNIA,CA,uoxp-street
+ehc-firstname,smemp-name,1950/04/10,wrv-city," 10080"," 93",974,FLORIDA,FL,wfrv-street
+vns-firstname,zmkad-name,1955/01/25,aoe-city," 10080"," 174",479,FLORIDA,FL,qgbd-street
+iva-firstname,fusnk-name,1982/07/03,kqd-city," 14660"," 10",748,FLORIDA,FL,zmbq-street
+jxu-firstname,nfdtl-name,1982/07/05,isx-city," 12980"," 166",997,CALIFORNIA,CA,dldc-street
