+++ integration-tests/gcp/datasets/0007-single-thread-golden.csv
@@ -1,0 +1,40 @@
+birthdate,city,firstName,housenr,id,lastName,state,stateCode,street,zipCode
+1982/07/05,isx-city,jxu-firstname," 166",997,nfdtl-name,CALIFORNIA,CA,dldc-street," 12980"
+1950/04/10,wrv-city,ehc-firstname," 93",974,smemp-name,FLORIDA,FL,wfrv-street," 10080"
+1970/06/19,bet-city,oyu-firstname," 155",919,papgo-name,CALIFORNIA,CA,otzt-street," 10280"
+1958/06/02,kty-city,etl-firstname," 67",912,mpkrp-name,CALIFORNIA,CA,cpug-street," 10560"
+1974/12/11,yux-city,nmc-firstname," 65",897,ymjve-name,FLORIDA,FL,ycqn-street," 19280"
+1962/07/05,kfm-city,nnx-firstname," 151",807,pvcck-name,CALIFORNIA,CA,aybu-street," 19780"
+1982/07/19,aed-city,eti-firstname," 156",765,dhcim-name,FLORIDA,FL,syxk-street," 17140"
+1974/07/01,pss-city,llu-firstname," 198",756,jdmdi-name,CALIFORNIA,CA,utrs-street," 18060"
+1982/07/03,kqd-city,iva-firstname," 10",748,fusnk-name,FLORIDA,FL,zmbq-street," 14660"
+1971/01/05,alm-city,vgc-firstname," 110",731,yprgp-name,CALIFORNIA,CA,uoxp-street," 14660"
+1958/08/07,hgh-city,hzl-firstname," 85",681,jikyx-name,FLORIDA,FL,gmqp-street," 13320"
+1950/05/06,ald-city,trc-firstname," 1",662,temdh-name,FLORIDA,FL,tapb-street," 11720"
+1966/05/06,xus-city,guv-firstname," 164",630,rgryb-name,FLORIDA,FL,plka-street," 11160"
+1966/09/03,mkm-city,ldq-firstname," 9",600,mdftt-name,CALIFORNIA,CA,rjkt-street," 11460"
+1958/05/29,cyx-city,niq-firstname," 167",561,avxuw-name,CALIFORNIA,CA,dxit-street," 13520"
+1955/01/25,aoe-city,vns-firstname," 174",479,zmkad-name,FLORIDA,FL,qgbd-street," 10080"
+1962/05/03,qwb-city,wlf-firstname," 22",422,fmrwo-name,CALIFORNIA,CA,stcz-street," 17820"
+1958/10/21,mqw-city,tum-firstname," 156",389,qxdyx-name,FLORIDA,FL,gndl-street," 18100"
+1974/02/27,ute-city,ofx-firstname," 111",346,yxqsn-name,FLORIDA,FL,wqka-street," 18680"
+1958/06/08,utn-city,qtw-firstname," 176",320,pwwep-name,FLORIDA,FL,sheh-street," 15160"
+1954/06/22,kqf-city,sfu-firstname," 132",308,ibfdt-name,CALIFORNIA,CA,xbwg-street," 11120"
+1958/09/02,akw-city,skq-firstname," 14",304,cqapx-name,CALIFORNIA,CA,xakx-street," 11080"
+1966/03/19,zwq-city,egp-firstname," 102",296,qrenr-name,FLORIDA,FL,fbwi-street," 16920"
+1954/12/19,rew-city,wpy-firstname," 115",272,qxjcn-name,CALIFORNIA,CA,tldv-street," 11260"
+1954/01/02,bgh-city,xhp-firstname," 53",269,gkiis-name,FLORIDA,FL,btrx-street," 14440"
+1974/07/05,dec-city,nlv-firstname," 125",212,nmccn-name,FLORIDA,FL,txyt-street," 11780"
+1950/09/02,anm-city,aaz-firstname," 197",202,avrpe-name,FLORIDA,FL,ddjz-street," 14000"
+1982/02/18,sek-city,kxv-firstname," 84",181,bxerk-name,CALIFORNIA,CA,ctwu-street," 14180"
+1974/10/10,gki-city,cwd-firstname," 12",178,ujfpc-name,FLORIDA,FL,sgiq-street," 13800"
+1958/03/16,uxf-city,ahc-firstname," 131",177,nltvw-name,CALIFORNIA,CA,fwsy-street," 19520"
+1974/09/16,trs-city,ttv-firstname," 187",135,eqley-name,CALIFORNIA,CA,ibff-street," 12740"
+1954/11/15,ocs-city,ovv-firstname," 169",125,ejdoc-name,FLORIDA,FL,bwhd-street," 14920"
+1966/07/08,zhc-city,kzs-firstname," 88",96,reeba-name,FLORIDA,FL,abmv-street," 13100"
+1954/09/23,lai-city,pvw-firstname," 171",70,aavpj-name,FLORIDA,FL,nygu-street," 13780"
+1962/07/05,ljc-city,ayq-firstname," 163",59,ieigg-name,FLORIDA,FL,dnjz-street," 12960"
+1974/08/12,hkm-city,rgh-firstname," 134",44,wzkjq-name,CALIFORNIA,CA,yixf-street," 19000"
+1982/07/14,nvt-city,zio-firstname," 91",26,rezku-name,CALIFORNIA,CA,wwkd-street," 19080"
+1978/07/16,rtw-city,ayl-firstname," 179",25,bwkoe-name,CALIFORNIA,CA,mzww-street," 18660"
+1954/03/28,egm-city,wha-firstname," 20",12,vnaov-name,CALIFORNIA,CA,hpep-street," 13120"
