+++ integration-tests/gcp/metadata/pipeline-run-configuration/Dataflow.json
@@ -29,5 +29,7 @@
   },
   "configurationVariables": [],
   "name": "Dataflow",
-  "description": ""
+  "description": "",
+  "executionInfoLocationName": "",
+  "dataProfile": ""
 }
