+++ plugins/engines/beam/src/main/java/org/apache/hop/beam/transforms/io/BeamInputMeta.java
@@ -44,8 +44,8 @@ import java.util.Map;
 
 @Transform(
     id = "BeamInput",
-    name = "Beam Input",
-    description = "Describes a Beam Input",
+    name = "Beam File Input",
+    description = "Reads one or more file using the Beam execution framework",
     image = "beam-input.svg",
     categoryDescription = "i18n:org.apache.hop.pipeline.transform:BaseTransform.Category.BigData",
     keywords = "i18n::BeamInputMeta.keyword",
