+++ plugins/engines/beam/src/main/resources/org/apache/hop/beam/transforms/io/messages/messages_en_US.properties
@@ -20,12 +20,12 @@
 BeamInputDialog.DialogTitle=Beam Input
 BeamInputDialog.InputLocation=Input location
 BeamInputDialog.FileDefinition=File definition to use
-BeamInputMeta.keyword=beam,input
+BeamInputMeta.keyword=beam,input,file,text,csv
 BeamOutputDialog.DialogTitle=Beam Output
 BeamOutputDialog.OutputLocation=Output location
 BeamOutputDialog.FilePrefix=File prefix
 BeamOutputDialog.Windowed=Windowed writes?
 BeamOutputDialog.FileDefinition=File definition to use
 BeamOutputDialog.FileSuffix=File suffix
-BeamOutputMeta.keyword=beam,output
+BeamOutputMeta.keyword=beam,output,file,text,csv
